\name{dfcomb-package}
\alias{dfcomb-package}
\alias{dfcomb}
\docType{package}
\title{
\packageTitle{dfcomb}
}
\description{
\packageDescription{dfcomb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dfcomb}
\packageIndices{dfcomb}
}
\author{
\packageAuthor{dfcomb}

Maintainer: \packageMaintainer{dfcomb}
}
\references{
Riviere MK, Yuan Y, Dubois F, Zohar S (2014). A Bayesian dose-finding design for drug combination clinical trials based on the logistic model. Pharm Stat, 13, 4:247-57.
}
\keyword{ package design survival }
