\name{dfcomb-package}
\alias{dfcomb-package}
\alias{dfcomb}
\docType{package}
\title{
Phase I and Phase I/II Adaptive Dose-Finding Design for Combination Studies
}
\description{
This package provides functions to run simulations, and to determine the MTD 
combination in on-going trials, in the cases of (1) combination of two
agents where both toxicity and efficacy are assumed to increase with the dose 
of each agent (e.q. cytotoxic agents), and (2) combination of two agents where
toxicity is increasing with the dose of each agent, but the efficacy of one 
agent is increasing whereas the efficacy of the other agent can plateaus (e.g. 
a cytotoxic agent with a molecularly targeted agent).
}
\details{
\tabular{ll}{
Package: \tab dfcomb\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2014-12-08\cr
License: \tab GPL-3\cr
}
}
\author{
Jacques-Henri Jourdan and Marie-Karelle Riviere-Jourdan <eldamjh@gmail.com>
}
\references{
Riviere, MK, Yuan, Y, Dubois, F, Zohar, S (2014). A Bayesian dose-finding design for drug combination clinical trials based on the logistic model. Pharm Stat, 13, 4:247-57.
Riviere, MK, Yuan, Y, Dubois, F, Zohar, S (2014). A Bayesian dose finding design for clinical trials combining a cytotoxic agent with a molecularly targeted agent. JRSS-C.
}
