\name{dfcomb-package}
\alias{dfcomb-package}
\alias{dfcomb}
\docType{package}
\title{
\packageTitle{dfcomb}
}
\description{
\packageDescription{dfcomb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dfcomb}
\packageIndices{dfcomb}
}
\author{
\packageAuthor{dfcomb}

Maintainer: \packageMaintainer{dfcomb}
}
\references{
Riviere, MK, Yuan, Y, Dubois, F, Zohar, S (2014). A Bayesian dose-finding design for drug combination clinical trials based on the logistic model. Pharm Stat, 13, 4:247-57.
Riviere, MK, Yuan, Y, Dubois, F, Zohar, S (2014). A Bayesian dose finding design for clinical trials combining a cytotoxic agent with a molecularly targeted agent. JRSS-C.
}
\keyword{ package design survival }
