\name{dfmta-package}
\alias{dfmta-package}
\alias{dfmta}
\docType{package}
\title{
Phase I/II Adaptive Dose-Finding Design for Combination Studies
}
\description{
This package provides functions to run simulations and to determine the MTD 
combination in on-going trials in the cases of (1) combination of two
cytotoxic agents, and (2) combination of a molecularly targeted agent with a
cytotoxic agent.
}
\details{
\tabular{ll}{
Package: \tab dfmta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-02\cr
License: \tab GPL-3\cr
}
}
\author{
Jacques-Henri Jourdan and Marie-Karelle Riviere-Jourdan <eldamjh@gmail.com>
}
\references{
Riviere, MK, Yuan, Y, Dubois, F, Zohar, S (2014). A Bayesian dose-finding design for drug combination clinical trials based on the logistic model. Pharm Stat, 13, 4:247-57.
Riviere, MK, Yuan, Y, Dubois, F, Zohar, S (2014). A Bayesian dose finding design for clinical trials combining a cytotoxic agent with a molecularly targeted agent. JRSS-C.
}
